Provide PiP button helper
