+++ Demo/Sources/Router/Router.swift
@@ -79,8 +79,6 @@ extension Router: PictureInPictureDelegate {
         print("--> will stop")
     }
 
-    // TODO: Still to address: Player not stopped when closing from PiP overlay, next playback resumes at the same location
-
     func pictureInPictureDidStop(_ pictureInPicture: PictureInPicture) {
         print("--> did stop")
     }
