+++ Sources/Player/UserInterface/View.swift
@@ -15,13 +15,28 @@ public extension View {
     ///   - Returns: A view that fills the given binding when the player's publisher emits an
     ///   event.
     @ViewBuilder
-    func onReceive<T>(player: Player?, assign keyPath: KeyPath<PlayerProperties, T>, to binding: Binding<T>) -> some View
-        where T: Equatable {
-            if let player {
-                onReceive(player.propertiesPublisher, assign: keyPath, to: binding)
+    func onReceive<T>(player: Player?, assign keyPath: KeyPath<PlayerProperties, T>, to binding: Binding<T>) -> some View where T: Equatable {
+        if let player {
+            onReceive(player.propertiesPublisher, assign: keyPath, to: binding)
+        }
+        else {
+            self
+        }
+    }
+}
+
+public extension View {
+    func onRelease(perform release: @escaping () -> Void) -> some View {
+        onAppear {
+            PictureInPicture.shared.stop()
+        }
+        .onDisappear {
+            if PictureInPicture.shared.isActive {
+                PictureInPicture.shared.release = release
             }
             else {
-                self
+                release()
             }
+        }
     }
 }
